BF: force "little" endianness while xor_byte
We are using xor_byte for fingerprint computation and need to
also handle bytes of different length.  When working on big-endian
platforms (such as the only one left s390x on debian) our assumptions
of alignment of bytes in such strings fail. For consistency - always
do it "little endian" way